requirements.txt
annotated-types
anyio
asttokens
blinker
cachetools
certifi
charset-normalizer
click
colorama
comm
contourpy
cycler
debugpy
decorator
distlib
dnspython
executing
filelock
Flask
Flask-Cors
fonttools 
fsspec
gensim
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
gradio_client
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
huggingface-hub
idna
imbalanced-learn
imblearn
ipykernel
ipython
itsdangerous
jedi
Jinja2
joblib
jupyter_client
jupyter_core
kiwisolver
lightgbm
MarkupSafe
matplotlib
matplotlib-inline
nest-asyncio
nltk
numpy
packaging
pandas
parso
pillow
platformdirs
prompt_toolkit
proto-plus
protobuf
psutil
pure-eval
pyasn1
pyasn1_modules
pydantic
pydantic_core
Pygments
pymongo
pyparsing
python-dateutil
python-dotenv
pytz
pywin32
PyYAML
pyzmq
regex
requests
rsa
scikit-learn
scipy
seaborn
six
smart-open
sniffio
stack-data
threadpoolctl
tornado
tqdm
traitlets
typing_extensions
tzdata
uritemplate
urllib3
virtualenv
wcwidth
websockets
Werkzeug
wrapt
xgboost
